refactor(scenarios): move scenario slice to top-level scenarios/ domain (phase 0d)
Aligns with labelbus precedent — scenarios is a cross-domain meta subsystem
(fixtures span inventory/sales/procurement/workflow), not an inventory feature.
`inventory.` SQL schema placement is separate from Go package placement.
