+++ api/domain/http/scenarios/scenarioapi/route.go
@@ -4,11 +4,11 @@ import (
 	"net/http"
 
 	"github.com/timmaaaz/ichor/api/sdk/http/mid"
-	"github.com/timmaaaz/ichor/app/domain/inventory/scenarioapp"
+	"github.com/timmaaaz/ichor/app/domain/scenarios/scenarioapp"
 	"github.com/timmaaaz/ichor/app/sdk/auth"
 	"github.com/timmaaaz/ichor/app/sdk/authclient"
 	"github.com/timmaaaz/ichor/business/domain/core/permissionsbus"
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus"
 	"github.com/timmaaaz/ichor/foundation/logger"
 	"github.com/timmaaaz/ichor/foundation/web"
 )
