+++ app/domain/scenarios/scenarioapp/model.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 
 	"github.com/timmaaaz/ichor/app/sdk/errs"
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus"
 	"github.com/timmaaaz/ichor/foundation/timeutil"
 )
