+++ app/domain/scenarios/scenarioapp/order.go
@@ -1,7 +1,7 @@
 package scenarioapp
 
 import (
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus"
 	"github.com/timmaaaz/ichor/business/sdk/order"
 )
