+++ business/domain/scenarios/scenariobus/event.go
