+++ business/domain/scenarios/scenariobus/filter.go
