+++ business/domain/scenarios/scenariobus/model.go
