+++ business/domain/scenarios/scenariobus/stores/scenariodb/order.go
@@ -3,7 +3,7 @@ package scenariodb
 import (
 	"fmt"
 
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus"
 	"github.com/timmaaaz/ichor/business/sdk/order"
 )
