+++ business/sdk/dbtest/dbtest.go
@@ -58,8 +58,8 @@ import (
 	"github.com/timmaaaz/ichor/business/domain/inventory/inventorytransactionbus/stores/inventorytransactiondb"
 	"github.com/timmaaaz/ichor/business/domain/labels/labelbus"
 	"github.com/timmaaaz/ichor/business/domain/labels/labelbus/stores/labeldb"
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus"
-	"github.com/timmaaaz/ichor/business/domain/inventory/scenariobus/stores/scenariodb"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus"
+	"github.com/timmaaaz/ichor/business/domain/scenarios/scenariobus/stores/scenariodb"
 	"github.com/timmaaaz/ichor/business/domain/inventory/cyclecountitembus"
 	"github.com/timmaaaz/ichor/business/domain/inventory/cyclecountitembus/stores/cyclecountitemdb"
 	"github.com/timmaaaz/ichor/business/domain/inventory/cyclecountsessionbus"
